% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grand_prix_scraper.R
\name{qualifying_scraper}
\alias{qualifying_scraper}
\title{Scrape the formula1.com website for qualifying results for each Grand Prix for a given year.}
\usage{
qualifying_scraper(year)
}
\arguments{
\item{year}{A numeric value.}
}
\value{
A dataframe.
}
\description{
Scrape the formula1.com website for qualifying results for each Grand Prix for a given year.
}
\examples{
year <- 2022
sprint_grid_scraper(year)
}
