% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grand_prix_scraper.R
\name{race_result_scraper}
\alias{race_result_scraper}
\title{Scrape the formula1.com website for race results for each Grand Prix for a given year.}
\usage{
race_result_scraper(year)
}
\arguments{
\item{year}{A numeric value.}
}
\value{
A dataframe.
}
\description{
Scrape the formula1.com website for race results for each Grand Prix for a given year.
}
\examples{
year <- 2022
race_result_scraper(year)
}
