% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grand_prix_scraper.R
\name{sprint_grid_scraper}
\alias{sprint_grid_scraper}
\title{Scrape the formula1.com website for sprint grids for each sprint weekend for a given year.}
\usage{
sprint_grid_scraper(year)
}
\arguments{
\item{year}{A numeric value.}
}
\value{
A dataframe.
}
\description{
Scrape the formula1.com website for sprint grids for each sprint weekend for a given year.
}
\examples{
year <- 2022
sprint_grid_scraper(year)
}
